% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignRTbyHydro.R
\name{alignRTbyHydro}
\alias{alignRTbyHydro}
\title{Aligning retention time using hydrophobicity index}
\usage{
alignRTbyHydro(dat1, dat2, datHydroIndex, includeLength = FALSE,
  method = c("hydrosequence", "hydro"))
}
\arguments{
\item{dat1}{A data frame of a spectrum library to be aligned to}

\item{dat2}{A data frame containing the spectra to be predicted}

\item{datHydroIndex}{a data frame containing all the hydrophobicity
index of the peptides in both dat1 and dat2}

\item{includeLength}{a logic value indicating if including the peptide
length as a training feature (default is False)}

\item{method}{a string character indicating the method of RT alignment.
One of "HydroSequence" (default) and "Hydro".}
}
\value{
a data frame of the dat2 with the newly aligned retention time
}
\description{
Aligning retention time using hydrophobicity index
}
\examples{
file1 <- paste(system.file("files",package="iSwathX"),"Lib2.txt",sep="/")
file2 <- paste(system.file("files",package="iSwathX"),"Lib3.txt",sep="/")
file3 <- paste(system.file("files",package="iSwathX"),"hydroIndex.txt",sep="/")
dat1 <- normalise(readLibFile(file1))
dat2 <- normalise(readLibFile(file2))
datHydroIndex <- read.delim2(file3,sep="\\t",header=TRUE,as.is=TRUE)
dat2 <- alignRTbyHydro(dat1, dat2, datHydroIndex,includeLength=FALSE)
}
