% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildSpectraLibPair.R
\name{buildSpectraLibPair}
\alias{buildSpectraLibPair}
\title{Build a spectra library by integrating a pair of spectrum libraries}
\usage{
buildSpectraLibPair(baseLib, extLib, hydroIndex, method = c("time", "hydro",
  "hydrosequence"), cutoff.size = 50, cutoff.r2 = 0.8,
  recalibrateRT = TRUE, includeLength = FALSE, labelBase = NA,
  labelAddon = NA, formatBase = c("peakview", "openswath", "skyline",
  "spectronaut"), formatExt = c("peakview", "openswath", "skyline",
  "spectronaut"), outputFormat = c("peakview", "openswath", "skyline",
  "spectronaut"), outputFile = "NewLib.txt", plot = FALSE, clean = TRUE,
  merge = TRUE, consolidateAccession = FALSE, ...)
}
\arguments{
\item{baseLib}{A base library data frame or file}

\item{extLib}{An external/addon library data frame or file}

\item{hydroIndex}{A data frame or file containing peptide hydrophobicity index}

\item{method}{A character string to specify the RT alignment method. One of "time" (default),
"hydro", "hydrosequence" can be selected}

\item{cutoff.size}{A number value specifying the mimum number of common
peptides between two libraries requried for training}

\item{cutoff.r2}{A number value specifying the mimum correlation coefficient
of retention time that this method requires.}

\item{recalibrateRT}{A logic value indicating if retention times of external/addon library
adjust/shift in case of poor correlation}

\item{includeLength}{A logic value indicating if include the peptide
length as a feature for predicting retention time. Only applicable when method is "hydro"}

\item{labelBase}{A character string to specify the labels of proteins from the base library}

\item{labelAddon}{A character string to specify the labels of proteins from addon library}

\item{formatBase}{A character string denoting the file format of base library file. One of
"peakview" (default), "openswath", "skyline" and "spectronaut".}

\item{formatExt}{A character string denoting the file format of addon library file. One of
"peakview" (default), "openswath", "skyline" and "spectronaut".}

\item{outputFormat}{A character string denoting the file format of the output integrated library. One of
"peakview" (default), "openswath", "skyline" and "spectronaut".}

\item{outputFile}{A character string to specify the spectra library created.}

\item{plot}{a logic value representing if plots during processing will be plotted or not}

\item{clean}{A logic value representing if the input librarues will be cleaned before integration.
Default value is True}

\item{merge}{A logic value representing if the output will be the merged library (default),
or the adjested add-on library}

\item{consolidateAccession}{A logic value representing if the protein accessions will be consolidated to the
base library in the integrated library. Default value is True.}
}
\value{
A data frame of integrated spectrum library, also saved in working directory. Library analysis plots saved in
working directory
}
\description{
Build a spectra library by integrating a pair of spectrum libraries
}
\examples{
libfiles <- paste(system.file("files",package="iSwathX"),
                             c("Lib2.txt","Lib3.txt"),sep="/")
Lib2_3 <- buildSpectraLibPair(libfiles[1], libfiles[2], outputFormat = "peakview", clean = T)
}
