% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canonicalFormat.R
\name{canonicalFormat}
\alias{canonicalFormat}
\title{Standardise a sprectrum library data frame}
\usage{
canonicalFormat(dat, format = c("peakview", "openswath", "skyline",
  "spectronaut"))
}
\arguments{
\item{dat}{a data frame of a spectrum library}

\item{format}{a character string, representing the format of the input
spectrum library. One of "peakview" (default), "openswath", "skyline" and "spectronaut"
#' @return a data frame of the library in canonical format}
}
\description{
Standardise a sprectrum library data frame
}
\examples{
file <- paste(system.file("files",package="iSwathX"),"Lib1.txt",sep="/")
dat <- read.delim2(file,sep="\\t",stringsAsFactor = FALSE, header=TRUE)
OR
file <- paste(system.file("files",package="iSwathX"),"Lib3_skyline.csv",sep="/")
dat <- read.delim2(file,sep=",",stringsAsFactor = FALSE, header=TRUE)

dat <- try(canonicalFormat(dat, format = "peakview"))
}
