% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeIntensityCor.R
\name{computeIntensityCor}
\alias{computeIntensityCor}
\title{Plot for ion intensity ranking correlation of two libraries}
\usage{
computeIntensityCor(dat1, dat2, method = "spearman", nomod = FALSE)
}
\arguments{
\item{dat1}{A data frame containing the first spectrum library}

\item{dat2}{A data frame containing the second spectrum library}

\item{method}{a character string indicating the method for calculating
correlation coefficient. Either of "spearman" (default) or "kendall".}
}
\value{
a boxplot of ion intensity correlation between the two libraries and training data set
}
\description{
Plot for ion intensity ranking correlation of two libraries
}
\examples{
file1 <- paste(system.file("files",package="iSwathX"),"Lib2.txt",sep="/")
file2 <- paste(system.file("files",package="iSwathX"),"Lib3.txt",sep="/")
dat1 <- normalise(readLibFile(file1))
dat2 <- normalise(readLibFile(file2))
list.intensitycor <- computeIntensityCor(dat1, dat2)
list.intensitycor[[1]]
}
