% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeRTTime.R
\name{computeRTTime}
\alias{computeRTTime}
\title{Plot for retention time correlation of two libraries}
\usage{
computeRTTime(dat1, dat2, label1 = "Base Lib", label2 = "External Lib",
  nomod = FALSE)
}
\arguments{
\item{dat1}{A data frame containing the first spectrum library}

\item{dat2}{A data frame containing the second spectrum library}

\item{label1}{a character string representing the x axis label for plotting}

\item{label2}{a character string representing the y axis label for plotting}

\item{nomod}{a logic value, representing if the modified peptides and its
fragment ions will be removed. FALSE (default) means not removing.}
}
\value{
RT correlation and RT residual plots will be saved
under current working directory. a list of training set dataframe, RTcor plot, RT residual plot, and RT correlation value.
}
\description{
Plot for retention time correlation of two libraries
}
\examples{
libfiles <- paste(system.file("files",package="iSwathX"),
   c("Lib2.txt","Lib3.txt"),sep="/")
datBaseLib <- readLibFile(libfiles[1])
datExtLib <- readLibFile(libfiles[2])
list.timecor <- computeRTTime(datBaseLib, datExtLib)
list.timecor[[2]]
list.timecor[[3]]
}
