% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consolidateAccession.R
\name{consolidateAccession}
\alias{consolidateAccession}
\title{Consolidate the protein accessions between two libraries}
\usage{
consolidateAccession(datBaseLib, datExtLib)
}
\arguments{
\item{datBaseLib}{a data frame for base library}

\item{datExtLib}{a data frame for external/addon library}
}
\value{
a data frame of the external library with newly consolidated
protein accessions based on the base library.
}
\description{
Consolidate the protein accessions between two libraries
}
\examples{
file1 <- paste(system.file("files",package="iSwathX"),"Lib1.txt",sep="/")
file2 <- paste(system.file("files",package="iSwathX"),"Lib2.txt",sep="/")
datBaseLib <- readLibFile(file1)
datExtLib <- readLibFile(file2)
dat <- consolidateAccession(datBaseLib, datExtLib)
}
