% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRTrain.R
\name{getRTrain}
\alias{getRTrain}
\title{Get the training set form retention time prediction}
\usage{
getRTrain(dat1, dat2, nomod = TRUE)
}
\arguments{
\item{dat1}{a data frame of the base library}

\item{dat2}{a data frame of the add-on library}

\item{nomod}{a logic value indicating if the modified peptides are included
when generating the training set, default value is T, i.e., no modified peptides
included}
}
\value{
a data frame contained the retention time training set
}
\description{
Get the training set form retention time prediction
}
\examples{
libfiles <- paste(system.file("files",package="iSwathX"),
   c("Lib2.txt","Lib3.txt"),sep="/")
datBaseLib <- readLibFile(libfiles[1], clean=TRUE, nomod=FALSE, nomc=FALSE)
datExtLib <- readLibFile(libfiles[2], clean=TRUE, nomod=FALSE, nomc=FALSE)
datRTrain <- getRTrain(datBaseLib, datExtLib, nomod=TRUE)
}
