% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isMissCleavaged.R
\name{isMissCleavaged}
\alias{isMissCleavaged}
\title{Check if a sequence has miss cleavages}
\usage{
isMissCleavaged(seq, enz = c("trypsin", "gluc", "chymotrypsin"))
}
\arguments{
\item{seq}{A character string representing the peptide sequence}

\item{enz}{A character string representing the enzyme which can be one of
"trypsin" (defalut), "gluc", or "chymotrypsin"}
}
\value{
a boolean value
}
\description{
Check if a sequence has miss cleavages
}
\examples{
isMissCleavaged("VSQGSKDPAEGDGAQPEETPR")
}
