% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise.R
\name{normalise}
\alias{normalise}
\title{Normalisation of relative ion intensity}
\usage{
normalise(dat, method = c("top", "total"))
}
\arguments{
\item{dat}{a data frame of a spectrum library}

\item{method}{a character string representing the normalisation method.
One of "top" (default) or "total".}
}
\value{
a data frame with normalised ion intensity
}
\description{
Normalisation of relative ion intensity
}
\examples{
file <- paste(system.file("files",package="iSwathX"),"Lib2.txt",sep="/")
dat <- normalise(readLibFile(file))
}
