% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseAccession.R
\name{parseAccession}
\alias{parseAccession}
\title{Extract a (list of)  Uniprot accession(s) from a string}
\usage{
parseAccession(proteinID, method = c("keep", "replace"))
}
\arguments{
\item{proteinID}{a character string of protein ID which contains Uniprot
accession(s). Two formats are accepted:  1/P0CG40 or sp|O75369|FLNB_HUMAN}

\item{method}{a character string representing the method, one of "keep"
or "replace"}
}
\value{
all Uniprot accessions separated by space
}
\description{
Extract a (list of)  Uniprot accession(s) from a string
}
\examples{
parseAccession("sp|O75369|FLNB_HUMAN")  ## O75369
parseAccession("5/Q9ULI2/Q7KZI7/Q9P0L2/P27448/DECOY_Q6ZMI3") #Q9ULI2 Q7KZI7 Q9P0L2 P27448 Q6ZMI3
}
