% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakviewFormat.R
\name{peakviewFormat}
\alias{peakviewFormat}
\title{Convert a data frame into Peak View library format}
\usage{
peakviewFormat(datlib, nodup = TRUE)
}
\arguments{
\item{datlib}{a data frame of a canonical spectrum library}

\item{nodup}{A logic value, indicating if remove duplicated
sprectrum (default)}
}
\value{
a data frame of a spectrum library in Peak View format
}
\description{
Convert a data frame into Peak View library format
}
\examples{
libfile <- paste(system.file("files",package="iSwathX"),"Lib2.txt",sep="/")
dat <- readLibFile(libfile)
dat <- peakviewFormat(dat)
}
