% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStats.R
\name{plotStats}
\alias{plotStats}
\title{Plot the statistical summary of the base and external libraries}
\usage{
plotStats(datBaseLib, datExtLib, ...)
}
\arguments{
\item{datBaseLib}{a data frame for base library}

\item{datExtLib}{a data frame for external/addon library}

\item{...}{arguments to pass onto cleanLib function}
}
\value{
a list of ggplot objects of statistical plots of the two libraries, also saved in working directory
}
\description{
Plot the statistical summary of the base and external libraries
}
\examples{
libfiles <- paste(system.file("files",package="iSwathX"),c("Lib2.txt","Lib3.txt"),sep="/")
datBaseLib <- readLibFile(libfiles[1])
datExtLib <- readLibFile(libfiles[2])
list.statplots <- plotStats(datBaseLib, datExtLib)
}
