% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictRT.R
\name{predictRT}
\alias{predictRT}
\title{Predicting retention time based on time}
\usage{
predictRT(datBaseCommPart, datExtLibCommPart, datExtLibNewPart,
  cutoff.size = 50, cutoff.r2 = 0.8, cutoff.rse = 3)
}
\arguments{
\item{datBaseCommPart}{a data frame of common precursors in the base library}

\item{datExtLibCommPart}{a data frame of common precursors in the
external library}

\item{datExtLibNewPart}{a data frame of new spectra in the external library}

\item{cutoff.size}{a number value specifying the mimum number of common
peptides between two libraries requried for training}

\item{cutoff.r2}{a number value specifying the mimum correlation coefficient
of retention time that this method requires.}
}
\value{
a data frame which has the same spectra as datExtLibNewPart but with
new predicted retention time
}
\description{
Predicting retention time based on time
}
\examples{
libfiles <- paste(system.file("files",package="iSwathX"),
                             c("Lib2.txt","Lib3.txt"),sep="/")
datBaseLib <- readLibFile(libfiles[1], clean=TRUE, nomod=FALSE, nomc=FALSE)
datExtLib <- readLibFile(libfiles[2], clean=TRUE, nomod=FALSE, nomc=FALSE)
datExtLibNewPart <- splitLib(datBaseLib, datExtLib, nomod=TRUE)[["ExtNew"]]
datExtLibNewPart <-
       predictRT(datBaseLib,datBaseLib,datExtLibNewPart)

}
