% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proteinOverlap.R
\name{proteinOverlap}
\alias{proteinOverlap}
\title{Get the overlap proteins between two libraries}
\usage{
proteinOverlap(datBaseLib, datExtLib, parseAcc = TRUE)
}
\arguments{
\item{datBaseLib}{a data frame of a base library}

\item{datExtLib}{a data frame of an addon library}

\item{parseAcc}{a logic value indicating if the protein accession to be parsed. Default value
is True.}
}
\value{
A vector of protein accession that the two libraries overlap
}
\description{
Get the overlap proteins between two libraries
}
\examples{
libfiles <- paste(system.file("files",package="iSwathX"),
					c("Lib2.txt","Lib3.txt"),sep="/")
datBaseLib <- readLibFile(libfiles[1])
datExtLib <- readLibFile(libfiles[2])
accs <- proteinOverlap(datBaseLib,datExtLib)
}
