% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLibFile.R
\name{readLibFile}
\alias{readLibFile}
\title{Load a spectrum library into a data frame}
\usage{
readLibFile(file, format = c("peakview", "openswath", "skyline",
  "spectronaut"), type = c("spectrum", "hydro"), clean = TRUE, ...)
}
\arguments{
\item{file}{A file of a spectrum library, in .txt or .csv format, can be .gz files.}

\item{format}{A character string denoting the file format. One of "peakview" (default), "openswath",
"skyline" and "spectronaut"}

\item{type}{A character string denoting the file type. One of "spectrum" (default) and "hydro"}

\item{clean}{A logic value, representing if the library will be cleaned.}

\item{...}{named arguments to be passed to the cleanLib function.}
}
\value{
a data frame of the library with cleaning process
}
\description{
Load a spectrum library into a data frame
}
\examples{
file <- paste(system.file("files",package="iSwathX"),"Lib1.txt",sep="/")
dat <- readLibFile(file)
}
