% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectModel.R
\name{selectModel}
\alias{selectModel}
\title{Model selection for time-base RT prediction using cross validation}
\usage{
selectModel(datRTrain, k = 5)
}
\arguments{
\item{datRTrain}{a data frame of retention time training set, must include
two columns: RT_detected.base and RT_detected.}

\item{k}{a numeric value specifying the fold number of the cross validation}
}
\value{
a model object based on the best RMSE average on cross validation
}
\description{
Model selection for time-base RT prediction using cross validation
}
\examples{
libfiles <- paste(system.file("files",package="iSwathX"),
                             c("Lib2.txt","Lib3.txt"),sep="/")
datBaseLib <- readLibFile(libfiles[1], clean=TRUE, nomod=FALSE, nomc=FALSE)
datExtLib <- readLibFile(libfiles[2], clean=TRUE, nomod=FALSE, nomc=FALSE)


datRTrain <- getRTrain(datBaseLib, datExtLib, nomod=FALSE)
bestmodel <- selectModel(datRTrain)

}
