% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skylineFormat.R
\name{skylineFormat}
\alias{skylineFormat}
\title{Convert a data frame into Skyline library format}
\usage{
skylineFormat(dat, nodup = TRUE)
}
\arguments{
\item{dat}{a data frame of a canonical spectrum library}

\item{nodup}{A logic value, indicating if remove duplicated
sprectrum (default)}
}
\value{
a data frame of a spectrum library in Skyline format
}
\description{
Convert a data frame into Skyline library format
}
\examples{
libfile <- paste(system.file("files",package="iSwathX"),"Lib2.txt",sep="/")
dat <- readLibFile(libfile)
dat <- skylineFormat(dat)
}
