% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectronautFormat.R
\name{spectronautFormat}
\alias{spectronautFormat}
\title{Convert a data frame into Spectronaut library format}
\usage{
spectronautFormat(dat, nodup = TRUE)
}
\arguments{
\item{dat}{a data frame of a canonical spectrum library}

\item{nodup}{A logic value, indicating if remove duplicated
sprectrum (default)}
}
\value{
a data frame of a spectrum library in Spectronaut format
}
\description{
Convert a data frame into Spectronaut library format
}
\examples{
libfile <- paste(system.file("files",package="iSwathX"),"Lib2.txt",sep="/")
dat <- readLibFile(libfile)
dat <- spectronautFormat(dat)
}
