% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitData.R
\name{splitData}
\alias{splitData}
\title{Splits data frame into arbitrary number of groups by rows randomly}
\usage{
splitData(dat, k = 5)
}
\arguments{
\item{dat}{The data frame to be split into groups}

\item{k}{a numeric value representing the number of groups}
}
\value{
a list of data frames randomly split into a number of groups
}
\description{
Splits data frame into arbitrary number of groups by rows randomly
}
