% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitLib.R
\name{splitLib}
\alias{splitLib}
\title{Split spectra libraries into common and new spectra}
\usage{
splitLib(datBaseLib, datExtLib, alignSpecies = NA, nomod = TRUE)
}
\arguments{
\item{datBaseLib}{a data frame for base library}

\item{datExtLib}{a data frame for external/addon library}

\item{alignSpecies}{a character string or NA (default) representing the
species of the alignment base in the base library}

\item{ignoreCharge}{a logic value indicating if using precursor charge
states as a splitting factor}
}
\value{
a list of data frames composed of BaseCommon, ExtCommon and
ExtNew, corresponding to the common spectra of base library, common
spectra of external library and new spectra of external library
}
\description{
Split spectra libraries into common and new spectra
}
\examples{
libfiles <- paste(system.file("files",package="iSwathX"),
                             c("Lib2.txt","Lib3.txt"),sep="/")
datBaseLib <- readLibFile(libfiles[1], clean=TRUE, nomod=FALSE, nomc=FALSE)
datExtLib <- readLibFile(libfiles[2], clean=TRUE, nomod=FALSE, nomc=FALSE)
list.datLibs <- splitLib(datBaseLib, datExtLib, nomod=FALSE)
}
